front-end/backend-flask/requirements.txt
Flask
pyodbc
python-dotenv
faiss-gpu
numpy
supervision
# pip install -r requirements.txt